2022-swen344-WebEngineering/db-ajg7654/requirements.txt
psycopg2==2.8.4
pyyaml==5.1
numpy==1.21.5